Prepravka decimal poeni

+++ backend/src/main/java/com/metropolitan/quiz/dto/QuizResultResponse.java
@@ -1,13 +1,14 @@
 package com.metropolitan.quiz.dto;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public record QuizResultResponse(
         Long quizId,
         String quizTitle,
         Long studentId,
-        Integer score,
-        Integer maxScore,
+        BigDecimal score,
+        BigDecimal maxScore,
         OffsetDateTime submittedAt
 ) {
 }
